:bug: replace ints with bigints

+++ models/staging/stg_sdcs__kills.sql
@@ -12,15 +12,15 @@ clean as (
 
         -- foreign_ids
         campaign_id,
-        initiator_id::int as initiator_id,
+        initiator_id::bigint as initiator_id,
         initiator_user_id,
-        initiator_group_id::int as initiator_group_id,
+        initiator_group_id::bigint as initiator_group_id,
         {{target.schema}}.int_or_null(split_part(initiator_name, '|', 1)) as initiator_unit_id,
-        target_id::int as target_id,
-        target_group_id::int as target_group_id,
+        target_id::bigint as target_id,
+        target_group_id::bigint as target_group_id,
         {{target.schema}}.int_or_null(split_part(target_name, '|', 1)) as target_unit_id,
         target_user_id,
-        weapon_id::int as weapon_id,
+        weapon_id::bigint as weapon_id,
 
         -- datetime
         time_created as datetime_created,
